app/requirements.txt
gunicorn
flask
flask_cors
google-genai